feat: refactor list result and detail in 2 routes
