+++ src/app/simulation/resultat/CardRcu.tsx
