+++ src/app/simulation/resultat/DebugButton.tsx
