+++ src/app/simulation/resultat/FranceRenovBlock.tsx
