+++ src/app/simulation/resultat/Isolation.tsx
