+++ src/app/simulation/resultat/JaugeCoutlImage.tsx
