+++ src/app/simulation/resultat/JaugeDifficulteImage.tsx
