+++ src/app/simulation/resultat/JaugeEnvironnementalImage.tsx
