+++ src/app/simulation/resultat/NouvelleSimulation.tsx
