+++ src/app/simulation/resultat/ShowIsolationImages.tsx
