+++ src/app/simulation/resultat/SyncStore.tsx
