+++ src/app/simulation/resultat/Usage.tsx
