+++ src/app/simulation/resultat/[idSolution]/page.tsx
@@ -1,17 +1,19 @@
+import assert from "assert";
 import { Base64 } from "js-base64";
 import { type Metadata } from "next";
 
 import { estGlobalementRenove, informationBatimentSchema } from "@/lib/common/domain/InformationBatiment";
+import { type SolutionAvecEnergieCoutAide } from "@/lib/common/domain/values/SolutionAvecEnergieCoutAide";
 import { type TravauxNiveauIsolation } from "@/lib/common/domain/values/TravauxNiveauIsolation";
 
 import { sharedMetadata } from "../../../shared-metadata";
 import { fetchSolutions } from "../helper";
 import { SyncStore } from "../SyncStore";
-import { Resultat } from "./Resultat";
+import { DetailSolution } from "./DetailSolution";
 
-const title = "Résultat simulation";
-const description = "Résultat simulation";
-const url = "/simulation/resultat";
+const title = "Détail solution";
+const description = "Détail solution";
+const url = "/simulation/resultat-detail";
 
 export const metadata: Metadata = {
   ...sharedMetadata,
@@ -29,18 +31,17 @@ export const metadata: Metadata = {
 };
 
 export type Props = {
-  searchParams: {
-    complet: "non" | "oui";
-    hash: string;
+  params: {
     idSolution: string;
+  };
+  searchParams: {
+    hash: string;
     travauxNiveauIsolation: TravauxNiveauIsolation;
   };
 };
 
-const ResultatsPage = async ({ searchParams }: Props) => {
+const ResultatsPage = async ({ params: { idSolution }, searchParams }: Props) => {
   if (!searchParams.hash) throw new Error("Le hash est manquant");
-
-  const complet = searchParams.complet === "oui";
 
   const unparsedFormData: unknown = JSON.parse(Base64.decode(searchParams.hash));
   const formData = informationBatimentSchema.safeParse(unparsedFormData);
@@ -58,23 +59,28 @@ const ResultatsPage = async ({ searchParams }: Props) => {
     ? "Aucun"
     : searchParams.travauxNiveauIsolation ?? "Global";
 
-  const { solutions, nbSolutions, isRcuEligible } = await fetchSolutions({
+  const { solutions } = await fetchSolutions({
     informationBatiment,
     travauxNiveauIsolation,
-    complet,
+    complet: true,
   });
+
+  let detailSolution: SolutionAvecEnergieCoutAide | null = null;
+
+  if (idSolution) {
+    detailSolution = solutions.find(s => s.id === idSolution) || null;
+  }
+
+  assert(detailSolution, "La solution n'a pas été trouvée");
 
   return (
     <>
       <SyncStore hash={searchParams.hash} />
 
-      <Resultat
-        informationBatiment={formData.data}
-        solutions={solutions}
-        isRcuEligible={isRcuEligible}
-        complet={complet}
+      <DetailSolution
+        solution={detailSolution}
+        informationBatiment={informationBatiment}
         travauxNiveauIsolation={travauxNiveauIsolation}
-        nbSolutions={nbSolutions}
       />
     </>
   );
