+++ src/app/simulation/resultat/helper.tsx
