+++ src/app/simulation/resultat/loading.tsx
